backend_db/backend/pyproject.toml
[project]
name = "selfdb-backend"
version = "0.5.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "schemathesis>=4.6",
    "email-validator>=2.3",
    "psycopg[binary,pool]>=3.2.0",
    "locust>=2.42",
    "pydantic-settings>=2.0.0",
    "pwdlib[bcrypt]",
    "PyJWT",
    "apscheduler>=3.10.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
